refactor(engine): split ModelStatus into enum + ModelPipelineResult
ModelStatus is now the outcome enum (OK/OVERFLOW/UNDERFLOW/COMPLETE/
REJECTED); ModelPipelineResult is the reused mutable holder returned from
ModelPipeline.transform, carrying status()/consumed()/produced(). Mirrors
the enum/result split each common-* pipeline will adopt.
